+++ e2e/cep47/installed.ts
@@ -95,14 +95,11 @@ const test = async () => {
             val.transform.WriteCLValue
           );
           const clValue = maybeCLValue.unwrap();
-          console.log(clValue);
           if (clValue && clValue instanceof CLMap) {
             const hash = clValue.get(
               CLValueBuilder.string("contract_package_hash")
             );
-            console.log(hash);
             const event = clValue.get(CLValueBuilder.string("event_type"));
-            console.log(event);
             if (hash && contractPackageHash.split(5) === hash.value()) {
               acc = [...acc, { name: event.value(), clValue }];
             }
@@ -110,10 +107,6 @@ const test = async () => {
         }
         return acc;
       }, []);
-
-      if (events.length) {
-        console.log("xxx", events);
-      }
     }
   });
 
@@ -143,7 +136,7 @@ const test = async () => {
     KEYS.publicKey,
     ["1"],
     [new Map([['number', 'one']])],
-    MINT_ONE_PAYMENT_AMOUNT!,
+    "ABC",
     KEYS.publicKey,
     [KEYS]
   );
