--- packages/cep47-client/src/constants.ts
@@ -1,6 +1,0 @@
-export enum CEP47Events {
-  MintOne = "cep47_mint_one",
-  TransferToken = "cep47_transfer_token",
-  BurnOne = "cep47_burn_one",
-  MetadataUpdate = 'cep47_metadata_update'
-}
